Moved generate_json_map to new module
