+++ src/techui_builder/__main__.py
@@ -12,6 +12,7 @@ logger_ = logging.getLogger(__name__)
 
 app = typer.Typer(
     pretty_exceptions_show_locals=False,
+    context_settings={"allow_interspersed_args": True},
     help="""
     A script for building Phoebus GUIs.
