+++ src/techui_builder/autofill.py
@@ -8,7 +8,7 @@ from lxml import objectify
 from lxml.etree import Element, SubElement, tostring
 from lxml.objectify import ObjectifiedElement, fromstring
 
-from techui_builder.builder import _get_action_group
+from techui_builder.generate_jsonmap import _get_action_group
 from techui_builder.models import Component
 from techui_builder.utils import read_bob
