+++ tests/test_builder.py
@@ -1,20 +1,11 @@
 import logging
-import os
 from io import StringIO
 from pathlib import Path
 from unittest.mock import MagicMock, Mock, mock_open, patch
 
 import pytest
-from lxml import objectify
 from phoebusgen.widget import ActionButton, Group
 from softioc.builder import ClearRecords, records
-
-from techui_builder.builder import (
-    JsonMap,
-    _get_action_group,  # type: ignore
-    _get_nav_tabs,  # type: ignore
-    _serialise_json_map,  # type: ignore
-)
 
 
 @pytest.mark.parametrize(
@@ -300,341 +291,3 @@ def test_create_screens_extra_p_does_not_exist(builder_with_setup, caplog):
 
     for log_output in caplog.records:
         assert "Extra prefix BAD-PV" in log_output.message
-
-
-def test_write_json_map_no_synoptic(builder):
-    with pytest.raises(FileNotFoundError):
-        builder.write_json_map(synoptic=Path("bad-synoptic.bob"))
-
-
-def test_write_json_map(builder):
-    test_map = JsonMap(
-        str(Path(__file__).parent.joinpath("test_files/test_bob.bob")), None
-    )
-
-    # We don't want cover _generate_json_map in this test
-    builder._generate_json_map = Mock(return_value=test_map)
-
-    # Make sure opis/ dir exists
-    if not Path.exists(builder._write_directory):
-        os.mkdir(builder._write_directory)
-
-    # We don't want to access the _serialise_json_map function in this test
-    with patch("techui_builder.builder._serialise_json_map") as mock_serialise_json_map:
-        mock_serialise_json_map.return_value = {"test": "test"}
-
-        builder.write_json_map(
-            synoptic=builder._write_directory.joinpath("index.bob"),
-            dest=builder._write_directory,
-        )
-
-    dest_path = builder._write_directory.joinpath("JsonMap.json")
-    assert Path.exists(dest_path)
-
-    if Path.exists(dest_path):
-        os.remove(dest_path)
-
-
-# We don't want to access the _get_action_group function in this test
-@patch("techui_builder.builder._get_action_group")
-def test_generate_json_map(
-    mock_get_action_group,
-    builder_with_test_files,
-    example_json_map,
-    test_files,
-):
-    screen_path, dest_path = test_files
-
-    mock_xml = objectify.Element("action")
-    mock_xml["file"] = "test_child_bob.bob"
-    mock_get_action_group.return_value = mock_xml
-    builder_with_test_files._get_component_label = Mock(
-        side_effect=["Display", "Detector"]
-    )
-
-    test_json_map = builder_with_test_files._generate_json_map(
-        screen_path.absolute(), dest_path
-    )
-
-    assert test_json_map == example_json_map
-
-
-# TODO: write this test
-def test_generate_json_map_embedded_screen(builder_with_test_files, example_json_map):
-    builder_with_test_files._get_component_label = Mock(
-        side_effect=[
-            "Display",
-            "Detector",
-            "Embedded Display",
-            "Embedded Display",
-            "Embedded Display",
-        ]
-    )
-
-    screen_path = Path("tests/test_files/test_bob_embedded.bob").absolute()
-    dest_path = Path("tests/test_files/")
-
-    example_json_map.file = "test_bob_embedded.bob"
-    example_json_map.children.append(
-        JsonMap(
-            "$(IOC)/pmacAxis.pvi.bob",
-            display_name="Embedded Display",
-            exists=False,
-            macros={"M": "$(M)", "P": "$(P)"},
-        )
-    )
-
-    test_json_map = builder_with_test_files._generate_json_map(screen_path, dest_path)
-
-    assert test_json_map == example_json_map
-
-
-def test_generate_json_map_nav_tabs(builder_with_test_files, example_json_map_root):
-    builder_with_test_files._get_component_label = Mock(
-        side_effect=["Display", "Tab1", "Tab2"]
-    )
-
-    screen_path = Path("tests/test_files/test_bob_navtabs.bob").absolute()
-    dest_path = Path("tests/test_files/")
-
-    example_json_map_root.file = "test_bob_navtabs.bob"
-    example_json_map_root.children.extend(
-        [
-            JsonMap(display_name="Tab1", file="tab1.bob", exists=False),
-            JsonMap(display_name="Tab2", file="tab2.bob", exists=False),
-        ]
-    )
-
-    test_json_map = builder_with_test_files._generate_json_map(screen_path, dest_path)
-
-    assert test_json_map == example_json_map_root
-
-
-def test_parse_display_name_with_name(builder):
-    """Test parse display name when <name> tag is present"""
-    display_name = builder._parse_display_name(
-        "<name>", Path("/path/to/filename.pvi.bob")
-    )
-    assert display_name == "<name>"
-
-
-def test_parse_display_name_from_filepath(builder):
-    """Test parse display name when only filepath is present"""
-    display_name = builder._parse_display_name(None, Path("/path/to/filename.pvi.bob"))
-    assert display_name == "filename"
-
-
-def test_parse_display_name_returns_none(builder):
-    """Test parse display ensures JSON displayName will return null otherwise"""
-    display_name = builder._parse_display_name(None, Path(""))
-
-    assert display_name is None
-
-
-def test_fix_names_json_map_recursive(builder, example_display_names_json):
-    """Test duplicate names are enumerated correctly for all children"""
-
-    test_display_names_json = JsonMap(
-        str(Path(__file__).parent.joinpath("test_files/test_bob.bob")), None
-    )
-
-    test_display_names_json_det1 = JsonMap(
-        "test_child_bob.bob", "Detector", macros={"P": "PV-DET-01"}, exists=False
-    )
-    test_display_names_json_det2 = JsonMap(
-        "test_child_bob.bob", "Detector", macros={"P": "PV-DET-02"}, exists=False
-    )
-    test_display_names_json_det3 = JsonMap(
-        "test_child_bob.bob", "Detector", macros={"P": "PV-DET-03"}, exists=False
-    )
-    test_display_names_json_det4 = JsonMap(
-        "test_child_bob.bob", "Detector", macros={"R": "NON-P-MACRO"}, exists=False
-    )
-    test_display_names_json_dev1 = JsonMap(
-        "test_child_bob.bob", "Device", macros={"P": "PV-DEV-01"}, exists=False
-    )
-    test_display_names_json_dev2 = JsonMap(
-        "test_child_bob.bob", "Device", macros={"P": "PV-DEV-02"}, exists=False
-    )
-    test_display_names_json = JsonMap("test_bob.bob", "Beamline")
-
-    test_display_names_json_dev1.children.append(test_display_names_json_det1)
-    test_display_names_json_dev1.children.append(test_display_names_json_det2)
-    test_display_names_json_dev2.children.append(test_display_names_json_det3)
-    test_display_names_json_dev2.children.append(test_display_names_json_det4)
-    test_display_names_json.children.append(test_display_names_json_dev1)
-    test_display_names_json.children.append(test_display_names_json_dev2)
-
-    builder._fix_names_json_map(test_display_names_json)
-
-    assert test_display_names_json == example_display_names_json
-
-
-# We don't want to access the _get_action_group function in this test
-@patch("techui_builder.builder._get_action_group")
-def test_generate_json_map_get_macros(
-    mock_get_action_group,
-    builder_with_test_files,
-    example_json_map,
-    test_files,
-):
-    screen_path, dest_path = test_files
-
-    # Set a custom macro to test against
-    example_json_map.children[0].macros = {"macro": "value"}
-
-    mock_xml = objectify.Element("action")
-    mock_xml["file"] = "test_child_bob.bob"
-    macros = objectify.SubElement(mock_xml, "macros")
-    # Set a macro to test
-    macros["macro"] = "value"
-    builder_with_test_files._get_component_label = Mock(
-        side_effect=["Display", "Detector"]
-    )
-    mock_get_action_group.return_value = mock_xml
-
-    test_json_map = builder_with_test_files._generate_json_map(screen_path, dest_path)
-    assert test_json_map == example_json_map
-
-
-def test_generate_json_map_xml_parse_error(builder_with_test_files, test_files):
-    screen_path = Path("tests/test_files/test_bob_bad.bob").absolute()
-    _, dest_path = test_files
-
-    test_json_map = builder_with_test_files._generate_json_map(screen_path, dest_path)
-
-    assert test_json_map.error.startswith("XML parse error:")
-
-
-@patch("techui_builder.builder._get_action_group")
-def test_generate_json_map_other_exception(
-    mock_get_action_group,
-    builder_with_test_files,
-    test_files,
-):
-    screen_path, dest_path = test_files
-
-    mock_get_action_group.side_effect = Exception("Some exception")
-    builder_with_test_files._get_component_label = Mock(
-        side_effect=["Display", "Detector"]
-    )
-
-    test_json_map = builder_with_test_files._generate_json_map(screen_path, dest_path)
-
-    assert test_json_map.error != ""
-
-
-def test_serialise_json_map(example_json_map):
-    json_ = _serialise_json_map(example_json_map)  # type: ignore
-
-    assert json_ == {
-        "file": "test_bob.bob",
-        "children": [
-            {"file": "test_child_bob.bob", "displayName": "Detector", "exists": False}
-        ],
-        "displayName": "Display",
-    }
-
-
-def test_get_action_group():
-    test_bob = objectify.parse("tests/test_files/test_bob.bob")
-
-    widget = test_bob.find(".//widget")
-    assert widget is not None
-
-    action_group = _get_action_group(widget)
-    assert action_group is not None
-
-
-def test_get_action_group_no_action_elements():
-    test_bob = objectify.parse("tests/test_files/test_bob.bob")
-
-    widget = test_bob.find(".//widget")
-    assert widget is not None
-
-    # Clear the actions element
-    widget.actions = objectify.ObjectifiedElement()
-
-    action_group = _get_action_group(widget)
-    assert action_group is None
-
-
-def test_get_action_group_no_actions_group(caplog):
-    # Use a blank xml element
-    widget = objectify.ObjectifiedElement()
-    # TODO: Do widgets always have a name attr, or _can_ it be empty??
-    widget.name = "Test"
-
-    with caplog.at_level(logging.ERROR):
-        _get_action_group(widget)
-
-    for log_output in caplog.records:
-        assert "Actions group not found" in log_output.message
-
-
-def test_get_component_label(builder_with_test_files):
-    display_name = builder_with_test_files._get_component_label(
-        "motor",
-        None,
-        None,
-    )
-    assert display_name == "Motor Stage"
-
-
-def test_get_component_label_child_labels(builder_with_test_files):
-    display_name = builder_with_test_files._get_component_label(
-        "X",
-        current_component_name="motor",
-        display_name="X",
-    )
-    assert display_name == "X1"
-
-
-def test_get_component_label_child_labels_with_name_already_pregenerated(
-    builder_with_test_files,
-):
-    display_name = builder_with_test_files._get_component_label(
-        "X1",
-        current_component_name="motor",
-        display_name="X",
-    )
-    assert display_name == "X1"
-
-
-def test_get_component_label_with_name_elem_invalid(
-    builder_with_test_files,
-):
-    display_name = builder_with_test_files._get_component_label(
-        "invalid_name",
-        current_component_name=None,
-        display_name="new_name",
-    )
-    assert display_name == "new_name"
-
-
-def test_get_component_label_with_current_component_name_invalid(
-    builder_with_test_files,
-):
-    display_name = builder_with_test_files._get_component_label(
-        "invalid_name",
-        current_component_name="invalid_name",
-        display_name="new_name",
-    )
-    assert display_name == "new_name"
-
-
-def test_get_nav_tabs(example_xml_navtabs_widget):
-    tabs_widget = _get_nav_tabs(example_xml_navtabs_widget)
-
-    assert isinstance(tabs_widget, list)
-
-
-def test_get_nav_tabs_no_tabs_group(caplog):
-    mock_navtabs = MagicMock(spec=objectify.ObjectifiedElement)
-    mock_navtabs.name = "no_tabs"
-
-    with caplog.at_level(logging.ERROR):
-        _get_nav_tabs(mock_navtabs)
-
-    for log_output in caplog.records:
-        assert "Tabs group not found" in log_output.message
